fix: 物件別利回りを実績ベースに統一し全体利回りと定義を一致させる
- AssetsAccordion: PropertyAsset に actualAnnualIncome/dataMonths を追加
- 物件別表面利回りを rent_payments 実績 ÷ 取得価格に変更（主役）
- 満室想定利回り（現在家賃×12÷取得価格）は参考値として控えめに併記
- データ蓄積中（0 or <12ヶ月）の場合は amber 色で注記を表示
- 全体表面利回りも同定義: rent_payments あれば実績収入、なければ精算ベースにフォールバック
- lib/format.ts・AssetsChart・report/page を formatYen に統一

+++ supabase/setup_acquisition_price.sql
@@ -11,7 +11,7 @@ ALTER TABLE properties ADD COLUMN IF NOT EXISTS acquisition_price BIGINT;
 --   年間家賃収入 = 在室(occupied)の部屋の rent_amount 合計 × 12
 UPDATE properties
 SET acquisition_price = (
-  SELECT (ROUND(SUM(r.rent_amount)::float * 12.0 / 0.08 / 10000))::bigint * 10000
+  SELECT (ROUND((SUM(r.rent_amount) * 12.0 / 0.08 / 10000)::numeric))::bigint * 10000
   FROM rooms r
   WHERE r.property_id = properties.id
     AND r.status = 'occupied'
@@ -31,7 +31,7 @@ SELECT
   SUM(r.rent_amount) FILTER (WHERE r.status = 'occupied') * 12 AS annual_rent,
   p.acquisition_price,
   ROUND(
-    SUM(r.rent_amount) FILTER (WHERE r.status = 'occupied')::float * 12.0 / p.acquisition_price * 100,
+    (SUM(r.rent_amount) FILTER (WHERE r.status = 'occupied') * 12.0 / p.acquisition_price * 100)::numeric,
     2
   ) AS yield_pct
 FROM properties p
